requirements.txt
functions-framework
pandas
numpy
openpyxl
google-cloud-storage
google-cloud-bigquery
pyarrow